Footer en Header refractored naar main.jsx: Contact en Favourites pagina als test gebruikt om de juiste layout terug te krijgen. Rest van pagina's moeten nog + overbodige code verwijderen
