+++ src/App.jsx
@@ -1,3 +1,4 @@
+// App.jsx
 import React from 'react';
 import './App.css';
 import Home from './pages/Home/Home';
